test(helpdesk): refactor fixtures and remove redundant skip markers
