+++ tests/e2e/helpdesk/conftest.py
@@ -135,3 +135,46 @@ async def async_created_form(async_mpt_ops, form_data):
 @pytest.fixture
 async def async_created_published_form(async_mpt_ops, async_created_form):
     return await async_mpt_ops.helpdesk.forms.publish(async_created_form.id)
+
+
+@pytest.fixture
+def parameter_data(short_uuid):
+    return {
+        "name": f"E2E Helpdesk Parameter {short_uuid}",
+        "description": "E2E Created Helpdesk Parameter",
+        "scope": "Form",
+        "phase": "Request",
+        "type": "SingleLineText",
+        "multiple": False,
+        "displayOrder": None,
+        "constraints": {
+            "required": False,
+            "readonly": False,
+            "hidden": False,
+        },
+        "options": {
+            "hintText": "E2E Parameter Hint Text",
+            "placeholderText": "E2E Parameter Placeholder Text",
+        },
+    }
+
+
+@pytest.fixture(scope="session")
+def invalid_parameter_id():
+    return "PAR-0000-0000"
+
+
+@pytest.fixture
+def created_parameter(mpt_ops, parameter_data):
+    with create_fixture_resource_and_delete(
+        mpt_ops.helpdesk.parameters, parameter_data
+    ) as parameter:
+        yield parameter
+
+
+@pytest.fixture
+async def async_created_parameter(async_mpt_ops, parameter_data):
+    async with async_create_fixture_resource_and_delete(
+        async_mpt_ops.helpdesk.parameters, parameter_data
+    ) as parameter:
+        yield parameter
