+++ tests/e2e/helpdesk/parameter_groups/parameters/test_sync_parameters.py
@@ -5,7 +5,7 @@ import pytest
 from mpt_api_client.exceptions import MPTAPIError
 from mpt_api_client.resources.helpdesk.parameter_group_parameters import ParameterGroupParameter
 
-pytestmark = [pytest.mark.flaky, pytest.mark.skip(reason="Unskip after MPT-19967 fixed")]
+pytestmark = [pytest.mark.flaky]
 
 
 def test_get_parameter_group_parameter(
