+++ tests/e2e/helpdesk/parameters/test_sync_parameters.py
@@ -5,7 +5,7 @@ import pytest
 from mpt_api_client.exceptions import MPTAPIError
 from mpt_api_client.resources.helpdesk.parameters import Parameter
 
-pytestmark = [pytest.mark.flaky, pytest.mark.skip(reason="Unskip after MPT-19967 fixed")]
+pytestmark = [pytest.mark.flaky]
 
 
 def test_get_parameter(mpt_ops, created_parameter):
